util funcs updated

+++ ctap/include/ctap_utils.h
@@ -1,4 +1,3 @@
-
 #ifndef CTAP_UTILS_H
 #define CTAP_UTILS_H
 
@@ -10,7 +9,7 @@ extern "C" {
 
 void print_hex(uint8_t* data, size_t size);
 
-uint64_t timestamp(void);
+int timestamp(void);
 
 #ifdef __cplusplus
 }
